requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
apted==1.0.3
async-timeout==5.0.1
attrs==24.2.0
black==24.10.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
cmake==3.31.1
deepspeed==0.9.5
Distance==0.1.3
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
google-auth==2.36.0
google-auth-oauthlib==0.4.6
grpcio==1.68.1
hjson==3.1.0
huggingface-hub==0.26.3
idna==3.10
importlib_metadata==8.5.0
isort==5.13.2
Jinja2==3.1.4
Levenshtein==0.20.9
lightning-utilities==0.11.9
lit==18.1.8
lxml==4.9.3
Markdown==3.7
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.2.1
ninja==1.11.1.2
numpy==1.22.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.535.161
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.4.127
nvitop==1.3.2
oauthlib==3.2.2
omegaconf==2.3.0
packaging==24.2
pathspec==0.12.1
Pillow==9.4.0
pip==24.3.1
platformdirs==4.3.6
propcache==0.2.1
protobuf==3.20.3
psutil==6.1.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==1.10.19
PyPDF2==2.11.1
pytorch-lightning==2.0.4
PyYAML==6.0.2
rapidfuzz==2.15.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.5
setuptools==75.6.0
six==1.16.0
sympy==1.13.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.5.0
timm==0.6.13
tokenizers==0.13.3
tomli==2.2.1
torch==2.0.1
torchmetrics==1.6.0
torchvision==0.15.2
tqdm==4.67.1
transformers==4.31.0
triton==2.0.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wheel==0.45.1
xformers==0.0.22
yarl==1.18.3
zipp==3.21.0
